# docs/swagger.yaml
basePath: /deployer
definitions:
  controller.Response:
    properties:
      Data:
        type: object
      OpCode:
        type: integer
      OpDesc:
        type: string
    type: object
  core.Condition:
    properties:
      LastTransitionTime:
        type: string
      Status:
        type: string
      Type:
        type: string
    type: object
  core.Metadata:
    properties:
      Annotations:
        additionalProperties:
          type: string
        type: object
      CreateTime:
        type: string
      Finalizers:
        items:
          type: string
        type: array
      Labels:
        additionalProperties:
          type: string
        type: object
      Name:
        type: string
      Namespace:
        type: string
      ResourceVersion:
        type: integer
      Uid:
        type: string
      UpdateTime:
        type: string
    type: object
  core.Status:
    properties:
      Conditions:
        items:
          $ref: '#/definitions/core.Condition'
        type: array
      Phase:
        type: string
    type: object
...
  v1.ConfigMapRef:
    properties:
      Name:
        type: string
      Namespace:
        type: string
    type: object
...
  v1.K8sConfig:
    properties:
      ApiVersion:
        type: string
      Kind:
        type: string
      Metadata:
        $ref: '#/definitions/core.Metadata'
        type: object
      Spec:
        $ref: '#/definitions/v1.K8sYaml'
        type: object
      Status:
        $ref: '#/definitions/core.Status'
        type: object
    type: object
  v1.K8sHostRef:
    properties:
      Label:
        additionalProperties:
          type: string
        type: object
      ValueFrom:
        $ref: '#/definitions/v1.ValueFrom'
        type: object
    type: object
  v1.K8sYaml:
    properties:
      Action:
        type: string
      Chrony:
        type: string
      Etcd:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
      ExLb:
        type: string
      GPU:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
      Harbor:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
      K8s-master:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
      K8s-worker:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
      K8s-worker-new:
        properties:
          Hosts:
            items:
              $ref: '#/definitions/v1.K8sHostRef'
            type: array
        type: object
    type: object
...
  v1.ValueFrom:
    properties:
      ConfigMapRef:
        $ref: '#/definitions/v1.ConfigMapRef'
        type: object
      HostRef:
        type: string
    type: object
...
info:
  contact: {}
  description: An example of gin
  license:
    name: MIT
    url: https://github.com/wujie1993/waves/blob/master/LICENSE
  termsOfService: https://github.com/wujie1993/waves
  title: Golang Gin API
  version: "1.0"
paths:
...
  /api/v1/namespaces/{namespace}/k8sconfig:
    get:
      consumes:
      - application/json
      parameters:
      - default: default
        description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  items:
                    $ref: '#/definitions/v1.K8sConfig'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取所有k8s集群
      tags:
      - K8sConfig
    post:
      consumes:
      - application/json
      parameters:
      - default: default
        description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: k8s集群信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.K8sConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/v1.K8sConfig'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 创建单个k8s集群配置
      tags:
      - K8sConfig
  /api/v1/namespaces/{namespace}/k8sconfig/{name}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: 集群名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/v1.K8sConfig'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 删除单个k8s集群配置
      tags:
      - K8sConfig
    get:
      consumes:
      - application/json
      parameters:
      - default: default
        description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: 集群名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/v1.K8sConfig'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 获取单个k8s集群配置
      tags:
      - K8sConfig
    put:
      consumes:
      - application/json
      parameters:
      - default: default
        description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: 集群名称
        in: path
        name: name
        required: true
        type: string
      - description: k8s集群信息
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/v1.K8sConfig'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.Response'
            - properties:
                Data:
                  $ref: '#/definitions/v1.K8sConfig'
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.Response'
      summary: 更新单个k8s集群配置
      tags:
      - K8sConfig
...
swagger: "2.0"
...